// packages/core/src/contracts/types.d.ts
export interface ChoiceQuestionDef {
    type: 'choice';
    instructions: string;
    criteria: Record<string, string | null>;
}
export interface ScoreQuestionDef {
    type: 'score';
    instructions: string;
    criteria: string[];
}
export interface NoulQuestionDef {
    type: 'noul';
    instructions: string;
}
export type QuestionDef = ChoiceQuestionDef | ScoreQuestionDef | NoulQuestionDef;